tools/scripts/public/pdp/requirements.txt
pylink-square==1.0.0
prodict>=0.8.18
cryptography==37.0.4
yubihsm
packaging
requests
pyyaml